[#2051] do not allow direct execution of @Util and @Catch methods
