+++ framework/src/play/utils/Java.java
@@ -1,23 +1,6 @@
 package play.utils;
 
 
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.FutureTask;
-
 import javassist.ClassPool;
 import javassist.CtClass;
 import javassist.bytecode.SourceFileAttribute;
@@ -28,10 +11,7 @@ import play.data.binding.Binder;
 import play.data.binding.ParamNode;
 import play.data.binding.RootParamNode;
 import play.exceptions.UnexpectedException;
-import play.mvc.After;
-import play.mvc.Before;
-import play.mvc.Finally;
-import play.mvc.With;
+import play.mvc.*;
 
 import java.io.*;
 import java.lang.annotation.Annotation;
@@ -42,8 +22,8 @@ import java.util.*;
 import java.util.concurrent.FutureTask;
 
 import static java.util.Collections.addAll;
-import static org.apache.commons.io.IOUtils.closeQuietly;
 import static java.util.Collections.sort;
+import static org.apache.commons.io.IOUtils.closeQuietly;
 
 /**
  * Java utils
@@ -112,28 +92,6 @@ public class Java {
             throw new RuntimeException(e);
         }
     }
-
-    /**
-     * Find the first public static method of a controller class
-     * @param name The method name
-     * @param clazz The class
-     * @return The method or null
-     */
-    public static Method findActionMethod(String name, Class clazz) {
-        while (!clazz.getName().equals("java.lang.Object")) {
-            for (Method m : clazz.getDeclaredMethods()) {
-                if (m.getName().equalsIgnoreCase(name) && Modifier.isPublic(m.getModifiers())) {
-                    // Check that it is not an interceptor
-                    if (!m.isAnnotationPresent(Before.class) && !m.isAnnotationPresent(After.class) && !m.isAnnotationPresent(Finally.class)) {
-                        return m;
-                    }
-                }
-            }
-            clazz = clazz.getSuperclass();
-        }
-        return null;
-    }
-
 
     /**
      * Invoke a static method
