+++ framework/test-src/play/mvc/ActionInvokerTest.java
@@ -1,13 +1,19 @@
 package play.mvc;
 
-import org.junit.*;
 import org.junit.Before;
+import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertSame;
+import java.lang.reflect.Method;
+
+import static org.junit.Assert.*;
 
 public class ActionInvokerTest {
     private Object[] noArgs = new Object[0];
+
+//    @org.junit.Before
+//    public void playBuilderBefore() {
+//        new PlayBuilder().build();
+//    }
 
     @Before
     public void setUp() throws Exception {
@@ -52,6 +58,31 @@ public class ActionInvokerTest {
         assertSame(controllerInstance, Http.Request.current().controllerInstance);
     }
 
+    @Test
+    public void testFindActionMethod() throws Exception {
+        assertNull(ActionInvoker.findActionMethod("notExistingMethod", ActionClass.class));
+
+        ensureNotActionMethod("privateMethod");
+        ensureNotActionMethod("beforeMethod");
+        ensureNotActionMethod("afterMethod");
+        ensureNotActionMethod("utilMethod");
+        ensureNotActionMethod("catchMethod");
+        ensureNotActionMethod("finallyMethod");
+
+        Method m = ActionInvoker.findActionMethod("actionMethod", ActionClass.class);
+        assertNotNull(m);
+        assertEquals("actionMethod", m.invoke( new ActionClass()));
+
+        //test that it works with subclassing
+        m = ActionInvoker.findActionMethod("actionMethod", ActionClassChild.class);
+        assertNotNull(m);
+        assertEquals("actionMethod", m.invoke( new ActionClassChild()));
+    }
+
+    private void ensureNotActionMethod(String name) throws NoSuchMethodException {
+        assertNull(ActionInvoker.findActionMethod(ActionClass.class.getDeclaredMethod(name).getName(), ActionClass.class));
+    }
+
     public static class TestController extends Controller {
         public static String staticJavaMethod() {
             return "static";
@@ -92,5 +123,45 @@ public class ActionInvokerTest {
         public static String traitMethod(Object that) {
             return "static-with-" + that;
         }
+    }
+
+    private static class ActionClass {
+
+        private static String privateMethod() {
+            return "private";
+        }
+
+
+        public static String actionMethod() {
+            return "actionMethod";
+        }
+
+        @play.mvc.Before
+        public static String beforeMethod() {
+            return "before";
+        }
+
+        @After
+        public static String afterMethod() {
+            return "after";
+        }
+
+        @Util
+        public static void utilMethod() {
+        }
+
+        @Catch
+        public static void catchMethod() {
+        }
+
+        @Finally
+        public static String finallyMethod() {
+            return "finally";
+        }
+
+    }
+
+    private static class ActionClassChild extends ActionClass {
+
     }
 }
